Upgrade to 2.0.21 and fix issues.
We were doing a too careful check to see if we were running the right
msi file. However, it takes some time to the msi file to get ready and
our check was verifying the presence of a particular string which wasn't
present at the time the check was performed. That unneeded check has
been removed.